Move logconfig tests to a separate file
